+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## unreleased
 
+- Casing - Configure characters to keep
 - CLI: Fix YAML output to preserve x-version number formatting
 
 ## [1.31.0] - 2026-04-12